+++ SmartLMS.Business/ICommunityService.cs
@@ -1,0 +1,19 @@
+using SmartLMS.Models;
+using System.Collections.Generic;
+using System.Threading.Tasks;
+
+namespace SmartLMS.Business;
+
+public interface ICommunityService
+{
+    // Forum
+    Task<IEnumerable<Post>> GetLatestPostsAsync(int count);
+    Task<Post?> GetPostByIdAsync(int id);
+    Task<Post> CreatePostAsync(Post post);
+    
+    // Events
+    Task<IEnumerable<CommunityEvent>> GetUpcomingEventsAsync();
+    
+    // Resources
+    Task<IEnumerable<CommunityResource>> GetRecentResourcesAsync();
+}